Remove transfers and transaction phases from prototype.
If we perform transfer work between opening and committing a
transaction rather than as part of the transaction, we don't need
a client-side commit (etc) phase, and we don't need to serialize
transfer origins or manifests.  This keeps a ton of
only-slightly-relevant stuff out of the prototype.
